+++ src/components/common/Divider.tsx
@@ -1,0 +1,3 @@
+export default function Divider() {
+  return <div className="my-2 h-px w-full bg-[#DBEBFE]" />;
+}
